Move signature fns to signatures mod
