+++ bindings_node/src/signatures.rs
@@ -5,8 +5,30 @@ use napi_derive::napi;
 use std::ops::Deref;
 use xmtp_id::associations::{
   unverified::{NewUnverifiedSmartContractWalletSignature, UnverifiedSignature},
-  AccountId,
+  verify_signed_with_public_context, AccountId,
 };
+
+#[napi]
+pub fn verify_signed_with_public_key(
+  signature_text: String,
+  signature_bytes: Uint8Array,
+  public_key: Uint8Array,
+) -> Result<()> {
+  let signature_bytes = signature_bytes.deref().to_vec();
+  let signature_bytes: [u8; 64] = signature_bytes
+    .try_into()
+    .map_err(|_| Error::from_reason("signature_bytes is not 64 bytes long."))?;
+
+  let public_key = public_key.deref().to_vec();
+  let public_key: [u8; 32] = public_key
+    .try_into()
+    .map_err(|_| Error::from_reason("public_key is not 32 bytes long."))?;
+
+  Ok(
+    verify_signed_with_public_context(signature_text, &signature_bytes, &public_key)
+      .map_err(ErrorWrapper::from)?,
+  )
+}
 
 #[napi]
 #[derive(Eq, Hash, PartialEq)]
@@ -166,5 +188,26 @@ impl Client {
     }
 
     Ok(())
+  }
+
+  #[napi]
+  pub fn sign_with_installation_key(&self, signature_text: String) -> Result<Uint8Array> {
+    let result = self
+      .inner_client()
+      .context()
+      .sign_with_public_context(signature_text)
+      .map_err(ErrorWrapper::from)?;
+
+    Ok(result.into())
+  }
+
+  #[napi]
+  pub fn verify_signed_with_installation_key(
+    &self,
+    signature_text: String,
+    signature_bytes: Uint8Array,
+  ) -> Result<()> {
+    let public_key = self.inner_client().installation_public_key();
+    verify_signed_with_public_key(signature_text, signature_bytes, public_key.into())
   }
 }
